requirements.txt
Pillow>=9.1.0
numpy>=1.21.0
streamlit>=1.20.0
